+++ src/solutions/problem42.rs
@@ -22,7 +22,7 @@ pub fn solution() -> SolutionResult {
     let path = Path::new("data/p042_words.txt");
     let contents = File::open(&path).read_to_string().ok().expect("Cannot read file");
     let names = contents[].split(',');
-    Ok(names.filter(|&x| is_triangular(word_value(x))).count() as i32)
+    Ok(names.filter(|&x| is_triangular(word_value(x))).count() as i64)
 }
 
 #[cfg(test)]
